Restructure pdf
