+++ packages/gitbook/src/app/middleware/(space)/~gitbook/pdf/layout.tsx
@@ -1,24 +1,9 @@
-import { CustomizationRootLayout } from '@/components/RootLayout';
-import { defaultCustomizationForSpace } from '@/lib/utils';
-
+import { PDFRootLayout } from '@/components/PDF';
 import { getV1ContextForPDF } from './pointer';
 
-/**
- * Layout to be used for the site root. It fetches the customization data for the
- * site or space and initializes the CustomizationRootLayout with it.
- */
-export default async function PDFRootLayout(props: { children: React.ReactNode }) {
+export default async function RootLayout(props: { children: React.ReactNode }) {
     const { children } = props;
-
     const context = await getV1ContextForPDF();
 
-    return (
-        <CustomizationRootLayout
-            customization={
-                'customization' in context ? context.customization : defaultCustomizationForSpace()
-            }
-        >
-            {children}
-        </CustomizationRootLayout>
-    );
+    return <PDFRootLayout context={context}>{children}</PDFRootLayout>;
 }
