+++ packages/gitbook/src/app/middleware/(space)/~gitbook/pdf/pointer.ts
@@ -5,23 +5,6 @@ import { getSiteContentPointer, getSpacePointer } from '@/lib/pointer';
 import { fetchV1ContextForSitePointer, fetchV1ContextForSpacePointer } from '@/lib/v1';
 
 /**
- * PDF generation can be done at the site level (e.g. docs.foo.com/~gitbook/pdf) or
- * at the space level (e.g. open.gitbook.com/~space/:spaceId/~gitbook/pdf) which is
- * for PDF export of in-app private spaces.
- *
- * This function returns the pointer depending on the context.
- */
-export async function getSiteOrSpacePointerForPDF(): Promise<
-    SiteContentPointer | SpaceContentPointer
-> {
-    try {
-        return await getSiteContentPointer();
-    } catch (_error) {
-        return getSpacePointer();
-    }
-}
-
-/**
  * Get the context for the PDF pointer.
  */
 export async function getV1ContextForPDF(): Promise<GitBookSiteContext | GitBookSpaceContext> {
@@ -30,4 +13,19 @@ export async function getV1ContextForPDF(): Promise<GitBookSiteContext | GitBook
     return 'siteId' in pointer && pointer.siteId
         ? await fetchV1ContextForSitePointer(pointer)
         : await fetchV1ContextForSpacePointer(pointer);
+}
+
+/**
+ * PDF generation can be done at the site level (e.g. docs.foo.com/~gitbook/pdf) or
+ * at the space level (e.g. open.gitbook.com/~space/:spaceId/~gitbook/pdf) which is
+ * for PDF export of in-app private spaces.
+ *
+ * This function returns the pointer depending on the context.
+ */
+async function getSiteOrSpacePointerForPDF(): Promise<SiteContentPointer | SpaceContentPointer> {
+    try {
+        return await getSiteContentPointer();
+    } catch (_error) {
+        return getSpacePointer();
+    }
 }
