+++ packages/gitbook/src/components/PDF/index.ts
@@ -1,0 +1,3 @@
+export * from './PDFRootLayout';
+export * from './PDFPage';
+export * from './urls';
